Changed the logic for packing
Now, it's not using a "random" pack name based on the first two
characters of the UUID, but a sequential integer ID, and reusing
a file until when its size goes above a threshold.

This is useful per-se, but is a preliminary needed step to
switch to a model where the key is a SHA sum, as the key (UUID vs. SHA)
cannot be known a priori, but only after the fact, so we cannot determine
the packing logic based on the key.