Fixing some failing tests
With the hashkey implementation it is not true anymore that the
number of objects I add results in the same number of objects being
written to disk (or to a pack) if I have the same object more than once.

Also, I am marking the test_concurrency test as an `xfail` until when
issue #4 is fixed.