Improving the windows logic in case of concurrent writes
This is for concurrent writes of loose objects.
I now raise if the conditions are really unexpected (there is a corrupted
loose object, I delete it, and while I try to recreate it, it reappears and
it is again corrupted). Note that if it reappears but it's OK, then
I assume everything went fine.

Also logic changed a bit to ensure that sandbox file is removed.

Now also testing the re-creation of content, either correct or corrupted,
using pytest's monkeypatching features.