Fixing issue in creation of pack files and committing to DB
Before, we were committing to the DB *before* closing the file.
This means that another concurrent process could find the information
on a packed file in the DB *before* the file was flushed and closed.

Now instead I flush, fsync it (so it's also much more robust in case
of unexpected failures, shutdowns, ...) and then, *after* closing,
I commit to the DB and delete the loose files.

This might address #4, to be checked.
For this, I am removing the xfail from that test.

Note that there might still be issues on Windows, with the loose file
being deleted while open - to check.