Limiting the maximum number of list elements in IN queries
Indeed, there is a limit and the default is small (999).
On Windows, it seems this default is unchanged (it's increased
in (some) Unix/Linux distributions and on Mac.
However, the safe approach is to always ask for this in chunks
(now of size 950).

Also fixing a 'performance bug' where some objects were returned
multiple times by the internal get_objects_stream functions.
Now I also test this.