Returning more metadata when iterating over objects
We now don't return only the size, but a set of metadata.
Note that this a backward-incompatible change (for those who
were using the internal get_objects_stream_and_meta (that anyway
used to be called get_objects_stream_and_size, and does not exist anymore).
Therefore, I upped the version to 0.4.0.

I also added a get_object_stream_and_meta (at the singular) for a single object.