Added seek method to both the PackedObjectReader and the StreamDecompresser
In this way, `seek` and `tell` work for both.

Note that only seeking with absolute position (using `whence=0`) is implemented.

Note that for the `StreamDecompresser`, only seeking to zero is efficient; any other
value will imply seeking back to zero and then decompressing the stream until to the
target position.
Improvements can be addressed later in #59