Adding methods to just fetch metas without opening streams
The methods are called `get_objects_meta` and `get_object_meta`.
The first just return a generator, the second is a standard function.
Since they don't open streams, they do not need to be context managers.

I therefore refactored the internal function `_get_objects_stream_meta_generator`
to try to keep the logic in a single place.

Also, I test that this method does not open new files, and I now use this
method for `has_objects`. Results oscillate a lot but seem to be consistently
better.

Therefore this fixes #48