Implement `seek(whence=1)`for `PackedObjectReader` and `StreamDecompresser`
This mode of seek is often used, for example by `numpy.load` and not
supporting it would prohibit a `Container` based repository with packed
objects from being used with it. Here we add support for `whence=1`
which simply means that the `target` is with respect to the current
byte position as reported by `tell` instead of with `0`.

Tests are not just added for the utility classes directly, but also on a
stream returned through `Container.get_object_stream`. This should
return an instance of `PackedObjectReader` if the object is packed and
an instance of `StreamDecompresser` if it was compressed when packed.
All scenarios are tested by parametrizing the test over the packing and
compression mode.