Fixing the locking mechanism in Windows
Now the basic tests should provide a valid locking mechanism,
that depends on specific libraries that exist only on Windows.
Dev dependencies have been updated accordingly.

I left over some stuff for locking in one of the tests because
it does not belong there, but should be moved to a different test
where we should mock os.rename