Fix to the hash calculation when writing to compressed packs.
The hasher was applied in the wrong place
(while writing rather than while reading) therefore it was hashing
*after* compression and storing with the wrong hash.
The hasher should run on the read stream instead.

We fix this and we also add tests.
We also add a validation while packing loose objects, so we avoid
to pack files that are corrupt (rasing an Exception).

Fixes #73