Objects were not retrieved in disk order anymore + bulk store
The main reason why the functions like `get_objects_stream*` do not
return objects in the same order is for efficiency, and this is achieved
by reorganising objects per pack, and reading the pack file in sequential
order.

During the code changes, this was not happening anymore.
Ensuring now that packed objects are returned in disk order (test also added)

Moreover, adding a lot of objects one to one is much less efficient than
doing a bulk insert with lower-level non-ORM functionality of SQLAlchemy.
For this reason, I am now committing in bulk all data to the DB at the end of each
pack file. Some comments on the performance in issue #92.