Reverting the logic of the export function
Now we define an `import_objects` function instead, that
(when importing from a container using the same hashing algorithm)
can be much more efficient by deciding beforehand which objects to
import, and then importing only those.
This uses the efficient method of #93 to iterate only once through two
sorted unique lists to get their intersection or left/right difference.

The performance still needs to be benchmarked.
The `export` function might be deprecated if this is more efficient.