Adding a new threshold to iterate on all objects when asking for too many hash keys
Above a certain threshold, it's just faster to iterate on all items rather than
making an SQL `IN` query on chunks of `_IN_SQL_MAX_LENGTH = 950` (e.g. if we have
millions of objects).
For now, this is set as `_MAX_CHUNK_ITERATE_LENGTH = 9500`, but this needs to be
benchmarked.

Also, this is still missing tests and a real benchmark.