Improve import time by moving `sqlalchemy` imports within function
This was discovered as a result of this library being used in
`aiida-core`, whose CLI command `verdi` increased in loading time
significantly once it started to import `disk_objectstore`.

Profiling pointed to `sqlalchemy` being the culprit. By moving the
imports from the module level to within the function scope, the loading
time is reduced from 0.11 seconds on average to about 0.03 seconds. This
reduction by 3 to 4 times is significant.