Update the repack handling
Since pack can sit in the additional folder, it no longer make sense
to get the path of a repack soley from the pack_id. Path to the
repack file must be obtained by giving the pack_id itself.
This ensures that the repack file sit on the same filessytem as the
pack that is being repacked, and allow safe handling at the end using
hard link.