only include part of the hashkey as filename
The full hashkey is 64 bytes long, that is a bit too much. The need
here is to ensure each file haev different names. If there are two
files with the same name the ZIP file is still valid, but requires
careful handling when extracting the file.

Optionally, we can verify filename uniqueness when sealing the pack.