Remove the writting of data descriptor
The data descriptor is designed for non-seekable stream. Previously,
it was used to store the file/compressed lengths. However, since
the pack file is only made to be valid ZIP file at the time of sealing
we can skip the datadescriptor and instead update the CRC, file_size
and compressed_size at the time of sealling, togther with the filename.

This save the 20 bytes data descriptor for each record.
Note that the CRC still needs to be stored. I make it store in the
SQLite database for now. This means that old SQLite database may
need to be migrated. Alternatiely, I can have it in a different table.
It can be decided later.