Add whence=2 to be used
For reading packed file as it is, there is no need to restrict
whence to be only 0 or 1.
Compressed files are more tricky as it is not possible to freely
seek to the end. Instead, the entire files will be decompressed back
into a loose file, which will then be opened for reading.
If such file exists already it will be used, so we don't decompress
twice. Such "cache" files are deleted during the routine maintainance
operations (e.g. pack_all_loose).