Refactoring the code to enable efficient access to packed compressed objects
I started to work on top of Bonan's work.
I kept the same core idea: upon certain conditions (now well defined, i.e.
when seeking with the following conditions):
- whence=0 and seeking in a previous location
- whence=1 and seeking in a previous location (i.e. negative offset)
- whence=2
we assume that we will need to do random access, so we just uncompress
the whole file back into the loose folder and then proxy all requests
(tell, seek, read) to the loose file.

I now define a LazyLooseStream class that allows to define which loose
file to open, delaying the opening to a later point, and in this way
enabling code that ensures that always closes any open file.

I also added code to ensure that there should not be race conditions
if a clean_storage is running at the same time.

I also cleaned up a bit the code and added various tests to increased coverage,
since it had dropped over time. It didn't go back to 100% but we are close
(for the core library files).

Furthermore, I used the occasion to a new `validate` CLI command
that also uses tqdm (if installed) to show progress.