Implement recompression in repacking, and auto compression heuristics
We implement here the feature to change the compression of each
pack when the pack is repacked. This gives the possibility to change
the compression if, when `pack_all_loose` was called, a different
choice was made.
This fixes #98.

In addition, we implement also a new compression mode,
`CompressMode.AUTO`, where some relatively inexpensive heuristics
are implemented to decide whether it's worth compressing or not,
at an objet level.
This fixes #14. Note that basic tests show that the function does
what it is supposed to do in simple cases - the heuristics can be
improved (in accuracy, or in speed) in the future, without any
backward-incompatible issue.

In doing so, also the parameter `compress` to `pack_all_loose`
accepts a `CompressMode` object, allowing to alrady use the
new `CompressMode.AUTO` mode.

Instead, the functions to write directly to packs only
support compression for all or no objects. The best
way is probably to not compress by default, and then
repack at the end with the `CompressMode.AUTO` mode
(this is anyway recommended, as there might be wasted space
if the same object is added multiple times to a pack directly,
and repacking will clean up this space).

In this PR we also fix a couple of minor GitHub Actions failing
tests (in benchmarks and periodic tests).