Reverting the sample_size and max_sampled_data_size
From my experiments, 32kB of sample size was too optimistic
(e.g. it estimated a compression to 88% on a video that instead,
if fully compressed was only 99.2% - the new 1kB sample size
returns a more accurate 99.5%.

On a long ~90MB log file it gives 9.1% instead of the real 3.8%,
but I think it's fine (it means in any case that I should compress).