Replacing `os.path` with `pathlib`
Note that in the process I changed the signature of some
of the functions to get (or return) only `pathlib.Path`
objects rather than strings, where appropriate.
Therefore, this PR is backward-incompatible and the next
release should be a new 0.7, or even directly 1.0.

However, for convenience, the main `Container.__init__`
method accepts both a string or a Path object, for convenience.

Fixes #149