Close sessions properly after usage
The session created during initialisation of the container was never
properly closed. This unclosed session was until py3.12 garbage
collected since it was unreferenced. With py3.13 the sessions however
are not anymore garbage collected and thus remain open. Resulting in
an open file descriptors of the `pack.idx` for each initialisation of
the container.

This commit fixes it by keeping track of the session that initialises
the container `_container_session`. We adapt the name `_session` to
`_operation_session` for a clearer distinguishment between the two
session types.