Close connection of database after initialization correctly (#179)
The session created during initialization of the container was never
properly closed. This unclosed session was until py3.12 garbage
collected since it was unreferenced. With py3.13 the sessions however
are not anymore garbage collected and thus remain open. Resulting in
an open file descriptors of the `pack.idx` for each initialization of
the container.

This commit fixes it by keeping track of the session that initializes
the container `_container_session`. We adapt the name `_session` to
`_operation_session` for a clearer distinction between the two
session types.