Fix "wrong" pack sizes due to missing flush by passing known size via `tell` (#200)
Earlier version was deciding if a pack file reached its target size by calling `stat()` on the file.
However, if a file is open, this might see a slightly smaller size due to buffering.
While this is not a major issue (typical buffer size is only a few kB), for correctness we instead pass the known file size to the function checking the pack size.
This, by the way, also improves efficiency.

Co-authored-by: Giovanni Pizzi <gio.piz@gmail.com>